# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for the EatMe recipe generation service",
        "title": "EatMe API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/profile": {
            "get": {
                "description": "Gets the profile for a user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get user profile",
                "operationId": "getProfile",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Profile"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    }
                }
            },
            "put": {
                "description": "Save a user's profile",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Save user profile",
                "operationId": "saveProfile",
                "parameters": [
                    {
                        "description": "User profile",
                        "name": "profile",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ProfileUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Profile"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    }
                }
            }
        },
...
        "/signup": {
            "post": {
                "description": "Register a new user account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Create a new user account",
                "operationId": "signup",
                "parameters": [
                    {
                        "description": "User signup information",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SignupRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SignupResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    },
                    "409": {
                        "description": "Email already exists",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.APIError": {
            "description": "APIError represents an API error response",
            "type": "object",
            "required": [
                "code",
                "message"
            ],
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {
                    "type": "string"
                },
                "field": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "models.Allergy": {
            "type": "string",
            "enum": [
                "dairy",
                "eggs",
                "fish",
                "gluten",
                "peanuts",
                "soy",
                "tree_nuts",
                "wheat"
            ],
            "x-enum-varnames": [
                "AllergyDairy",
                "AllergyEggs",
                "AllergyFish",
                "AllergyGluten",
                "AllergyPeanuts",
                "AllergySoy",
                "AllergyTreeNuts",
                "AllergyWheat"
            ]
        },
...
        "models.Cuisine": {
            "type": "string",
            "enum": [
                "american",
                "british",
                "chinese",
                "french",
                "german",
                "indian",
                "italian",
                "japanese",
                "mexican",
                "spanish",
                "thai",
                "vietnamese"
            ],
            "x-enum-varnames": [
                "CuisineAmerican",
                "CuisineBritish",
                "CuisineChinese",
                "CuisineFrench",
                "CuisineGerman",
                "CuisineIndian",
                "CuisineItalian",
                "CuisineJapanese",
                "CuisineMexican",
                "CuisineSpanish",
                "CuisineThai",
                "CuisineVietnamese"
            ]
        },
        "models.Diet": {
            "type": "string",
            "enum": [
                "vegetarian",
                "vegan",
                "keto",
                "paleo",
                "low_carb",
                "high_protein"
            ],
            "x-enum-varnames": [
                "DietVegetarian",
                "DietVegan",
                "DietKeto",
                "DietPaleo",
                "DietLowCarb",
                "DietHighProtein"
            ]
        },
        "models.Equipment": {
            "type": "string",
            "enum": [
                "stove",
                "oven",
                "microwave",
                "toaster",
                "grill",
                "smoker",
                "slow_cooker",
                "pressure_cooker",
                "sous_vide"
            ],
            "x-enum-varnames": [
                "EquipmentStove",
                "EquipmentOven",
                "EquipmentMicrowave",
                "EquipmentToaster",
                "EquipmentGrill",
                "EquipmentSmoker",
                "EquipmentSlowCooker",
                "EquipmentPressureCooker",
                "EquipmentSousVide"
            ]
        },
...
        "models.Profile": {
            "description": "Profile represents a user's profile information",
            "type": "object",
            "required": [
                "allergies",
                "cuisines",
                "diet",
                "equipment",
                "name",
                "setup_step",
                "skill"
            ],
            "properties": {
                "allergies": {
                    "description": "User's allergies",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Allergy"
                    }
                },
                "cuisines": {
                    "description": "User's cuisines",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Cuisine"
                    }
                },
                "diet": {
                    "description": "User's diet restrictions",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Diet"
                    }
                },
                "equipment": {
                    "description": "User's equipment",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Equipment"
                    }
                },
                "name": {
                    "description": "User's name",
                    "type": "string"
                },
                "setup_step": {
                    "description": "Setup Step",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.SetupStep"
                        }
                    ]
                },
                "skill": {
                    "description": "User's skill level",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Skill"
                        }
                    ]
                }
            }
        },
        "models.ProfileUpdateRequest": {
            "description": "ProfileUpdateRequest represents a user's profile update request payload",
            "type": "object",
            "properties": {
                "allergies": {
                    "description": "User's allergies",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Allergy"
                    }
                },
                "cuisines": {
                    "description": "User's cuisines",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Cuisine"
                    }
                },
                "diet": {
                    "description": "User's diet restrictions",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Diet"
                    }
                },
                "equipment": {
                    "description": "User's equipment",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Equipment"
                    }
                },
                "name": {
                    "description": "User's name",
                    "type": "string"
                },
                "setup_step": {
                    "description": "Setup Step",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.SetupStep"
                        }
                    ]
                },
                "skill": {
                    "description": "User's skill level",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Skill"
                        }
                    ]
                }
            }
        },
...
        "models.SetupStep": {
            "type": "string",
            "enum": [
                "profile",
                "skill",
                "cuisines",
                "diet",
                "equipment",
                "allergies",
                "done"
            ],
            "x-enum-varnames": [
                "SetupStepProfile",
                "SetupStepSkill",
                "SetupStepCuisines",
                "SetupStepDiet",
                "SetupStepEquipment",
                "SetupStepAllergies",
                "SetupStepDone"
            ]
        },
        "models.SignupRequest": {
            "description": "SignupRequest represents the user signup request payload",
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "description": "User's email address",
                    "type": "string",
                    "example": "john.doe@example.com"
                },
                "password": {
                    "description": "User's password",
                    "type": "string",
                    "example": "Password123!"
                }
            }
        },
        "models.SignupResponse": {
            "description": "SignupResponse represents the user signup response",
            "type": "object",
            "required": [
                "token"
            ],
            "properties": {
                "token": {
                    "description": "Access token for user",
                    "type": "string",
                    "example": "\u003cJWT_TOKEN\u003e"
                }
            }
        },
        "models.Skill": {
            "type": "string",
            "enum": [
                "beginner",
                "intermediate",
                "advanced",
                "chef"
            ],
            "x-enum-varnames": [
                "SkillBeginner",
                "SkillIntermediate",
                "SkillAdvanced",
                "SkillChef"
            ]
        },
...
    }
}